requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
EasyProcess==1.1
email_validator==2.1.1
entrypoint2==1.1
fastapi==0.111.0
fastapi-cli==0.0.3
fonttools==4.45.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.5
networkx==3.2.1
numpy==1.24.3
orjson==3.10.3
packaging==23.2
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pydantic==2.7.1
pydantic_core==2.18.2
PyGetWindow==0.0.9
Pygments==2.18.0
pymongo==4.7.2
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
simpy==4.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.11.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2023.3
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
